feat(api-service): update sync use new issue creation
