+++ apps/api/src/app/bridge/bridge.module.ts
@@ -19,18 +19,10 @@ import { CommunityOrganizationRepository, PreferencesRepository } from '@novu/da
 import { SharedModule } from '../shared/shared.module';
 import { BridgeController } from './bridge.controller';
 import { USECASES } from './usecases';
-import { PostProcessWorkflowUpdate } from '../workflows-v2/usecases/post-process-workflow-update';
-import { OverloadContentDataOnWorkflowUseCase } from '../workflows-v2/usecases/overload-content-data';
-import {
-  BuildDefaultPayloadUsecase,
-  CollectPlaceholderWithDefaultsUsecase,
-  PrepareAndValidateContentUsecase,
-  ValidatePlaceholderUsecase,
-} from '../workflows-v2/usecases/validate-content';
 import { BuildAvailableVariableSchemaUsecase } from '../workflows-v2/usecases/build-variable-schema';
-import { ExtractDefaultValuesFromSchemaUsecase } from '../workflows-v2/usecases/extract-default-values-from-schema';
 import { HydrateEmailSchemaUseCase } from '../environments-v1/usecases/output-renderers/hydrate-email-schema.usecase';
 import { BuildPayloadSchema } from '../workflows-v2/usecases/build-payload-schema/build-payload-schema.usecase';
+import { BuildStepIssuesUsecase } from '../workflows-v2/usecases/build-step-issues/build-step-issues.usecase';
 
 const PROVIDERS = [
   CreateWorkflow,
@@ -48,18 +40,12 @@ const PROVIDERS = [
   UpsertPreferences,
   DeletePreferencesUseCase,
   UpsertControlValuesUseCase,
-  PostProcessWorkflowUpdate,
-  OverloadContentDataOnWorkflowUseCase,
-  PrepareAndValidateContentUsecase,
   BuildAvailableVariableSchemaUsecase,
-  BuildDefaultPayloadUsecase,
-  ValidatePlaceholderUsecase,
-  CollectPlaceholderWithDefaultsUsecase,
-  ExtractDefaultValuesFromSchemaUsecase,
   TierRestrictionsValidateUsecase,
   HydrateEmailSchemaUseCase,
   CommunityOrganizationRepository,
   BuildPayloadSchema,
+  BuildStepIssuesUsecase,
 ];
 
 @Module({
