+++ apps/api/src/app/workflows-v2/shared/compute-workflow-status.ts
@@ -1,0 +1,19 @@
+import { StepIssues, WorkflowStatusEnum } from '@novu/shared';
+import { NotificationStep } from '@novu/application-generic';
+
+export function computeWorkflowStatus(workflowActive: boolean, steps: NotificationStep[]) {
+  if (!workflowActive) {
+    return WorkflowStatusEnum.INACTIVE;
+  }
+
+  const hasIssues = steps.filter((step) => hasControlIssues(step.issues)).length > 0;
+  if (!hasIssues) {
+    return WorkflowStatusEnum.ACTIVE;
+  }
+
+  return WorkflowStatusEnum.ERROR;
+}
+
+export function hasControlIssues(issue: StepIssues | undefined) {
+  return issue?.controls && Object.keys(issue.controls).length > 0;
+}
