--- apps/api/src/app/workflows-v2/usecases/extract-default-values-from-schema/extract-default-values-from-schema.command.ts
@@ -1,7 +1,0 @@
-import { BaseCommand } from '@novu/application-generic';
-import { JSONSchemaDto } from '@novu/shared';
-
-export class ExtractDefaultValuesFromSchemaCommand extends BaseCommand {
-  jsonSchemaDto?: JSONSchemaDto;
-  controlValues: Record<string, unknown>;
-}
