--- apps/api/src/app/workflows-v2/usecases/extract-default-values-from-schema/index.ts
@@ -1,2 +1,0 @@
-export * from './extract-default-values-from-schema.command';
-export * from './extract-default-values-from-schema.usecase';
