+++ apps/api/src/app/workflows-v2/usecases/index.ts
@@ -1,11 +1,8 @@
 export * from './build-step-data';
 export * from './build-test-data';
 export * from './build-variable-schema';
-export * from './extract-default-values-from-schema';
 export * from './generate-preview';
 export * from './get-workflow';
 export * from './list-workflows';
 export * from './sync-to-environment';
 export * from './upsert-workflow';
-export * from './validate-content';
-export * from './post-process-workflow-update';
