--- apps/api/src/app/workflows-v2/usecases/overload-content-data/index.ts
@@ -1,2 +1,0 @@
-export * from './overload-content-data-on-workflow.usecase';
-export * from './overload-content-data-on-workflow.command';
