--- apps/api/src/app/workflows-v2/usecases/overload-content-data/overload-content-data-on-workflow.command.ts
@@ -1,5 +1,0 @@
-import { EnvironmentWithUserObjectCommand, WorkflowInternalResponseDto } from '@novu/application-generic';
-
-export class OverloadContentDataOnWorkflowCommand extends EnvironmentWithUserObjectCommand {
-  workflow: WorkflowInternalResponseDto;
-}
