--- apps/api/src/app/workflows-v2/usecases/post-process-workflow-update/index.ts
@@ -1,2 +1,0 @@
-export * from './post-process-workflow-update.command';
-export * from './post-process-workflow-update.usecase';
