--- apps/api/src/app/workflows-v2/usecases/post-process-workflow-update/post-process-workflow-update.command.ts
@@ -1,5 +1,0 @@
-import { EnvironmentWithUserObjectCommand, WorkflowInternalResponseDto } from '@novu/application-generic';
-
-export class PostProcessWorkflowUpdateCommand extends EnvironmentWithUserObjectCommand {
-  workflow: WorkflowInternalResponseDto;
-}
