--- apps/api/src/app/workflows-v2/usecases/validate-content/build-payload-from-placeholder/build-default-payload.command.ts
@@ -1,6 +1,0 @@
-import { BaseCommand } from '@novu/application-generic';
-import { ValidatedPlaceholderAggregation } from '../validate-placeholders';
-
-export class BuildDefaultPayloadCommand extends BaseCommand {
-  placeholderAggregators: ValidatedPlaceholderAggregation[];
-}
