--- apps/api/src/app/workflows-v2/usecases/validate-content/build-payload-from-placeholder/build-default-payload.response.ts
@@ -1,5 +1,0 @@
-import { PreviewPayload } from '@novu/shared';
-
-export class BuildDefaultPayloadResponse {
-  previewPayload: PreviewPayload;
-}
