--- apps/api/src/app/workflows-v2/usecases/validate-content/build-payload-from-placeholder/index.ts
@@ -1,3 +1,0 @@
-export * from './build-default-payload.response';
-export * from './build-default-payload.command';
-export * from './build-default-payload.usecase';
