--- apps/api/src/app/workflows-v2/usecases/validate-content/collect-placeholders/collect-placeholder-with-defaults.command.ts
@@ -1,5 +1,0 @@
-import { BaseCommand } from '@novu/application-generic';
-
-export class CollectPlaceholderWithDefaultsCommand extends BaseCommand {
-  controlValues?: Record<string, unknown>;
-}
