--- apps/api/src/app/workflows-v2/usecases/validate-content/collect-placeholders/index.ts
@@ -1,3 +1,0 @@
-export * from './collect-placeholder-with-defaults.usecase';
-export * from './placeholder.aggregation';
-export * from './collect-placeholder-with-defaults.command';
