--- apps/api/src/app/workflows-v2/usecases/validate-content/collect-placeholders/placeholder.aggregation.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export interface PlaceholderAggregation {
-  nestedForPlaceholders: Record<string, Record<string, string>>;
-  regularPlaceholdersToDefaultValue: Record<string, string>;
-}
