--- apps/api/src/app/workflows-v2/usecases/validate-content/prepare-and-validate-content/index.ts
@@ -1,3 +1,0 @@
-export * from './prepare-and-validate-content.command';
-export * from './prepare-and-validate-content.usecase';
-export * from './validated-content.response';
