--- apps/api/src/app/workflows-v2/usecases/validate-content/prepare-and-validate-content/prepare-and-validate-content.command.ts
@@ -1,11 +1,0 @@
-import { EnvironmentWithUserObjectCommand } from '@novu/application-generic';
-import { JSONSchemaDto, PreviewPayload, StepTypeEnum } from '@novu/shared';
-
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export interface PrepareAndValidateContentCommand extends EnvironmentWithUserObjectCommand {
-  controlValues: Record<string, unknown>;
-  controlDataSchema: JSONSchemaDto;
-  variableSchema: JSONSchemaDto;
-  previewPayloadFromDto?: PreviewPayload;
-  stepType?: StepTypeEnum;
-}
