--- apps/api/src/app/workflows-v2/usecases/validate-content/prepare-and-validate-content/validated-content.response.ts
@@ -1,9 +1,0 @@
-// Define the ValidatedContent interface
-import { ContentIssue, PreviewPayload } from '@novu/shared';
-
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export interface ValidatedContentResponse {
-  finalPayload: PreviewPayload;
-  finalControlValues: Record<string, unknown>;
-  issues: Record<string, ContentIssue[]>;
-}
