--- apps/api/src/app/workflows-v2/usecases/validate-content/validate-placeholders/index.ts
@@ -1,3 +1,0 @@
-export * from './validate-placeholder.usecase';
-export * from './validate-placeholder.command';
-export * from './validated-placeholder-aggregation';
