--- apps/api/src/app/workflows-v2/usecases/validate-content/validate-placeholders/validate-placeholder.command.ts
@@ -1,8 +1,0 @@
-import { JSONSchemaDto } from '@novu/shared';
-import { PlaceholderAggregation } from '../collect-placeholders/placeholder.aggregation';
-
-// eslint-disable-next-line @typescript-eslint/naming-convention
-export interface ValidatePlaceholderCommand {
-  controlValueToPlaceholders: Record<string, PlaceholderAggregation>;
-  variableSchema: JSONSchemaDto;
-}
